moved commands to cmd/
