+++ src/cmd/commands.h
